+++ SSYMachaphore.h
@@ -29,7 +29,7 @@
 	NSString* name ;
 	NSPort* m_port ;
 	NSTimeInterval initialBackoff ;
-	float backoffFactor ;
+	CGFloat backoffFactor ;
 	NSTimeInterval maxBackoff ;
 	NSTimeInterval timeout ;
 }
@@ -37,7 +37,7 @@
 @property (copy) NSString* name ;
 @property (retain) NSPort* port ;
 @property NSTimeInterval initialBackoff ;
-@property float backoffFactor ;
+@property CGFloat backoffFactor ;
 @property NSTimeInterval maxBackoff ;
 @property NSTimeInterval timeout ;
 
@@ -67,7 +67,7 @@
  */
 - (void)setName:(NSString*)name_
  initialBackoff:(NSTimeInterval)initialBackoff_
-  backoffFactor:(float)backoffFactor_
+  backoffFactor:(CGFloat)backoffFactor_
 	 maxBackoff:(NSTimeInterval)maxBackoff_
 		timeout:(NSTimeInterval)timeout_ ;
